refactor(dashboards): address PR review on tile filter merge
- extract the pure filter-override precedence helpers out of InsightDetails.tsx
  into insightDetailsFilterOverrides.ts (file had grown too large)
- fix propertyIdentity to compare object-valued keys by value and to include
  group_type_index, matching the backend _property_identity so display
  attribution can't diverge from the computed merge
- dedupe the property value renderer in CompactUniversalFiltersDisplay
- correct the tile filters help text and the TileOverridesWarning tooltip to
  reflect that properties AND-combine while date/breakdown replace
- trim descriptive comments

Generated-By: PostHog Code
Task-Id: 1c680a43-4a6c-4f7d-b752-ef42ab907b02
