+++ frontend/src/lib/components/Cards/InsightCard/InsightDetails.tsx
@@ -68,11 +68,20 @@ import {
     isTrendsQuery,
     hasBreakdownFilter,
 } from '~/queries/utils'
-import { AnyPropertyFilter, BaseMathType, FilterLogicalOperator, PropertyGroupFilter, UserBasicType } from '~/types'
+import { AnyPropertyFilter, BaseMathType, FilterLogicalOperator, UserBasicType } from '~/types'
 
 import { PropertyKeyInfo } from '../../PropertyKeyInfo'
 import { TZLabel } from '../../TZLabel'
 import { CompactUniversalFiltersDisplay } from './CompactUniversalFiltersDisplay'
+import {
+    dropDuplicatesOfOverrides,
+    EffectiveDateOverride,
+    getDateRangeOverrideDisplay,
+    getEffectiveFilterOverrides,
+    OverrideSource,
+    PropertiesInput,
+    splitOutOverrideProperties,
+} from './insightDetailsFilterOverrides'
 
 export function InsightDetailSectionDisplay({
     icon,
@@ -96,123 +105,6 @@ export function InsightDetailSectionDisplay({
 
 function assertNever(value: never): never {
     throw new Error(`Unexpected entity node: ${(value as { kind?: string } | undefined)?.kind ?? 'unknown'}`)
-}
-
-type OverrideSource = 'dashboard' | 'tile'
-
-interface EffectiveFilterOverrides {
-    // Non-overlapping keys from both layers contribute; dashboard first to match backend order.
-    propertyGroups: { properties: AnyPropertyFilter[]; source: OverrideSource }[]
-    // Dashboard property filters the tile shadows on the same key — they don't apply, but we surface them
-    // struck-through so the precedence is visible rather than silently dropped.
-    overriddenByTile: AnyPropertyFilter[]
-    // Breakdown is a single value: tile wins when set, otherwise the dashboard's.
-    breakdown: { breakdownFilter: NonNullable<DashboardFilter['breakdown_filter']>; source: OverrideSource } | null
-}
-
-// NOTE: this re-derives the precedence encoded in the backend `merge_filters_by_priority` /
-// `remove_query_properties_overridden_by` purely to attribute each shown filter to its source
-// ("Dashboard"/"Tile"). It must stay in step with that backend rule — if the tie-break there changes
-// (e.g. a field is added to `_SCALAR_OVERRIDE_FIELDS`), update this too or the tags will mislead.
-
-// The (type, key) a property filter targets — the unit at which a tile takes precedence.
-// Mirrors backend `_property_identity`.
-function propertyIdentity(property: AnyPropertyFilter): string {
-    const type = 'type' in property ? property.type : 'event'
-    const key = 'key' in property ? property.key : ''
-    return `${type}::${key}`
-}
-
-// Tile and dashboard overrides merge per field (matches backend `merge_filters_by_priority`).
-// Property filters merge per key: a tile filter replaces the dashboard's on the same key.
-export function getEffectiveFilterOverrides(
-    filtersOverride: DashboardFilter | undefined,
-    tileFiltersOverride: TileFilters | null | undefined
-): EffectiveFilterOverrides {
-    const tileProperties = tileFiltersOverride?.properties ?? []
-    const tileKeys = new Set(tileProperties.map(propertyIdentity))
-    const dashboardProperties: AnyPropertyFilter[] = []
-    const overriddenByTile: AnyPropertyFilter[] = []
-    for (const property of filtersOverride?.properties ?? []) {
-        if (tileKeys.has(propertyIdentity(property))) {
-            overriddenByTile.push(property)
-        } else {
-            dashboardProperties.push(property)
-        }
-    }
-    const propertyGroups: EffectiveFilterOverrides['propertyGroups'] = []
-    if (dashboardProperties.length > 0) {
-        propertyGroups.push({ properties: dashboardProperties, source: 'dashboard' })
-    }
-    if (tileProperties.length > 0) {
-        propertyGroups.push({ properties: tileProperties, source: 'tile' })
-    }
-
-    const breakdown = tileFiltersOverride?.breakdown_filter
-        ? { breakdownFilter: tileFiltersOverride.breakdown_filter, source: 'tile' as const }
-        : filtersOverride?.breakdown_filter
-          ? { breakdownFilter: filtersOverride.breakdown_filter, source: 'dashboard' as const }
-          : null
-
-    return { propertyGroups, overriddenByTile, breakdown }
-}
-
-interface DateRangeSource {
-    date_from?: string | null
-    date_to?: string | null
-}
-
-interface EffectiveDateOverride {
-    source: OverrideSource
-    dateFrom: string | null | undefined
-    dateTo: string | null | undefined
-    // The lower-priority layer this replaced, shown struck-through so the precedence is visible.
-    replaced?: {
-        source: OverrideSource | 'insight'
-        dateFrom: string | null | undefined
-        dateTo: string | null | undefined
-    }
-}
-
-function hasDateBound(source: DateRangeSource | null | undefined): boolean {
-    // `!= null` (not truthiness) matches the backend's `is not None`, so an explicit empty-string bound counts.
-    return source?.date_from != null || source?.date_to != null
-}
-
-// Works out which layer's date range actually applies and what it replaced, so the popup can show the
-// tile/dashboard override with the overridden range struck-through. Mirrors the backend precedence in
-// `merge_filters_by_priority` (tile beats dashboard beats the insight's own range); returns null when the
-// insight's own range wins (nothing was overridden).
-export function getDateRangeOverrideDisplay(
-    insightDateRange: DateRangeSource | undefined,
-    filtersOverride: DashboardFilter | undefined,
-    tileFiltersOverride: TileFilters | null | undefined
-): EffectiveDateOverride | null {
-    let winner: {
-        source: OverrideSource
-        dateFrom: string | null | undefined
-        dateTo: string | null | undefined
-    } | null = null
-    if (hasDateBound(tileFiltersOverride)) {
-        winner = { source: 'tile', dateFrom: tileFiltersOverride?.date_from, dateTo: tileFiltersOverride?.date_to }
-    } else if (hasDateBound(filtersOverride)) {
-        winner = { source: 'dashboard', dateFrom: filtersOverride?.date_from, dateTo: filtersOverride?.date_to }
-    }
-    if (!winner) {
-        return null
-    }
-
-    let replaced: EffectiveDateOverride['replaced']
-    if (winner.source === 'tile' && hasDateBound(filtersOverride)) {
-        replaced = { source: 'dashboard', dateFrom: filtersOverride?.date_from, dateTo: filtersOverride?.date_to }
-    } else if (hasDateBound(insightDateRange)) {
-        replaced = { source: 'insight', dateFrom: insightDateRange?.date_from, dateTo: insightDateRange?.date_to }
-    }
-    if (replaced && replaced.dateFrom === winner.dateFrom && replaced.dateTo === winner.dateTo) {
-        replaced = undefined
-    }
-
-    return { ...winner, replaced }
 }
 
 const LAYER_LABELS: Record<OverrideSource | 'insight', string> = {
@@ -251,85 +143,6 @@ function OverrideNote({
             <span className="text-muted-alt">{children}</span>
         </div>
     )
-}
-
-// Normalize a filter value into a canonical key for comparison. A scalar and its single-element array
-// form mean the same filter (the insight editor and the dashboard bar store the same value differently),
-// and the value list is a set — order and duplicates don't matter — so compare as a sorted set.
-function normalizeFilterValue(value: unknown): string {
-    const entries = value == null ? [] : Array.isArray(value) ? value : [value]
-    return JSON.stringify([...new Set(entries.map((entry) => JSON.stringify(entry)))].sort())
-}
-
-// The override round-trips through the backend into the merged query and picks up normalized fields
-// the raw override lacks, so a deep-equal fails — compare on the fields that actually identify a filter.
-function isSamePropertyFilter(a: AnyPropertyFilter, b: AnyPropertyFilter): boolean {
-    const operatorOf = (f: AnyPropertyFilter): string | undefined => ('operator' in f ? f.operator : undefined)
-    return (
-        (a.type ?? 'event') === (b.type ?? 'event') &&
-        a.key === b.key &&
-        (operatorOf(a) ?? 'exact') === (operatorOf(b) ?? 'exact') &&
-        normalizeFilterValue(a.value) === normalizeFilterValue(b.value)
-    )
-}
-
-function samePropertyFilters(a: AnyPropertyFilter[], b: AnyPropertyFilter[]): boolean {
-    return a.length === b.length && a.every((f, i) => isSamePropertyFilter(f, b[i]))
-}
-
-// Matches the shape `convertPropertiesToPropertyGroup` accepts: a group, a flat list, or nothing.
-type PropertiesInput = PropertyGroupFilter | AnyPropertyFilter[] | null | undefined
-
-// Drop base leaves that are exact duplicates of a filter shown in a higher-priority override layer, so a
-// filter the insight and an override both set isn't listed twice — it shows once, on the layer that took
-// priority. Only exact matches (same type/key/operator/value) are collapsed; a shared key with a different
-// value is left alone, since both genuinely AND together.
-export function dropDuplicatesOfOverrides(
-    base: PropertiesInput,
-    overrideProperties: AnyPropertyFilter[]
-): PropertiesInput {
-    if (!base || overrideProperties.length === 0) {
-        return base
-    }
-    const isDuplicate = (leaf: AnyPropertyFilter): boolean =>
-        overrideProperties.some((override) => isSamePropertyFilter(leaf, override))
-    if (Array.isArray(base)) {
-        return base.filter((leaf) => !isDuplicate(leaf))
-    }
-    const values = (base.values ?? [])
-        .map((subgroup) =>
-            'values' in subgroup && Array.isArray(subgroup.values)
-                ? { ...subgroup, values: (subgroup.values as AnyPropertyFilter[]).filter((leaf) => !isDuplicate(leaf)) }
-                : subgroup
-        )
-        .filter((subgroup) => !('values' in subgroup && Array.isArray(subgroup.values) && subgroup.values.length === 0))
-    return { ...base, values }
-}
-
-// The query returned for a dashboard tile already has the override's properties ANDed in (as the
-// trailing subgroup/tail), so pull that part out to attribute it rather than list it twice.
-function splitOutOverrideProperties(
-    properties: PropertiesInput,
-    overrideProperties: AnyPropertyFilter[]
-): { base: PropertiesInput; overrideFound: boolean } {
-    if (!properties || overrideProperties.length === 0) {
-        return { base: properties, overrideFound: false }
-    }
-    // Flat list: the backend concatenated the override onto the end.
-    if (Array.isArray(properties)) {
-        const tailStart = properties.length - overrideProperties.length
-        if (tailStart >= 0 && samePropertyFilters(properties.slice(tailStart), overrideProperties)) {
-            return { base: properties.slice(0, tailStart), overrideFound: true }
-        }
-        return { base: properties, overrideFound: false }
-    }
-    // Group: the backend AND-wrapped the insight's group with the override as the final subgroup.
-    const subgroups = properties.values ?? []
-    const last = subgroups[subgroups.length - 1]
-    if (last && samePropertyFilters(last.values as AnyPropertyFilter[], overrideProperties)) {
-        return { base: { ...properties, values: subgroups.slice(0, -1) }, overrideFound: true }
-    }
-    return { base: properties, overrideFound: false }
 }
 
 function EntityDisplay({ entity }: { entity: AnyEntityNode<AnyDataWarehouseNode> }): JSX.Element {
@@ -612,19 +425,12 @@ export function PropertiesSummary({
     const overriddenProperties = overriddenByTile ?? []
     const allOverrideProperties = overrideGroups.flatMap((group) => group.properties)
     const { base, overrideFound } = splitOutOverrideProperties(properties, allOverrideProperties)
-    // A filter the insight and an override both set would otherwise show twice — collapse the base copy so
-    // it appears once, on the layer that took priority.
     const dedupedBase = overrideFound ? dropDuplicatesOfOverrides(base, allOverrideProperties) : base
-    // "Active filters" when overrides stack on top, so it's clear these are what actually applies vs the
-    // replaced ones shown below; plain "Filters" otherwise.
     const label = overrideFound ? 'Active filters' : 'Filters'
     return (
         <InsightDetailSectionDisplay icon={<IconFilter />} label={label}>
-            {/* Label the base as the insight's own only when overrides stack on top, so the layers read
-                as a clear insight → dashboard → tile stack. Plain insights stay unlabeled. */}
             {overrideFound && <OverrideNote source="insight">base filters:</OverrideNote>}
             <CompactUniversalFiltersDisplay groupFilter={convertPropertiesToPropertyGroup(dedupedBase)} />
-            {/* overrideFound means we removed the overrides from the list above, so show them once here. */}
             {overrideFound &&
                 overrideGroups.map((group) => (
                     <React.Fragment key={group.source}>
@@ -634,8 +440,7 @@ export function PropertiesSummary({
                         />
                     </React.Fragment>
                 ))}
-            {/* Dashboard filters the tile shadowed — labelled with the same layer badges as the applied
-                rows ([Dashboard] replaced by [Tile]) and struck/dimmed so it reads as no longer applied. */}
+            {/* Dashboard filters the tile shadowed — shown struck-through so precedence is visible. */}
             {overrideFound && overriddenProperties.length > 0 && (
                 <>
                     <LemonDivider className="my-2" />
